Use onos topo service commands to add devices to running clusters.
